test(cli): start splitting cli test suite
